Move nRF52 Hardware Crypto To Extra File
...in order to be reused somewhere else.
