+++ examples/nrf52/src/hw_crypto.h
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2022 Felix Gohla, Konrad Hanff, Tobias Kantusch,
+ *                    Quentin Kuth, Felix Roth. All rights reserved.
+ *
+ * Use of this source code is governed by a BSD-style
+ * license that can be found in the LICENSE file.
+ */
+
+/**
+ * @brief Initialize cryptography module (if hardware cryptography is enabled).
+ * 
+ * @return int 0 on success.
+ */
+int init_hw_crypto();
